*GPT working for all supported FS types.
git-svn-id: https://wiiflow-advanced.googlecode.com/svn/trunk@325 43c7226f-5572-e80b-0314-b0280708705f

+++ source/loader/utils.h
@@ -28,7 +28,7 @@ extern "C" {
 #endif	/* __cplusplus */
 
 /* Prototypes */
-u32 le64(u64);
+u64 le64(u64);
 u32 le32(u32);
 u16 le16(u16);
 int makedir(char *newdir);
